# src/docs/openapi.yml
openapi: 3.0.3
info:
  title: FaceMatch 企業級管理系統 API
  description: |
    FaceMatch 承攬商管理系統的完整 REST API 文檔
    
    ## 功能特色
    - 🔐 企業級 AD 驗證 + 本地帳號支援
    - 👥 承攬商管理 (完整 CRUD)
    - 📋 施工單管理 + 多層級簽核工作流程
    - 🎓 年度資格管理 + 快速操作
    - 👤 使用者管理 + 角色權限控制
    - 📊 企業級操作日誌追蹤
    - ⚡ FaceMatch 整合管理
    
    ## 認證方式
    使用 Bearer Token 進行 API 認證
    
    ## API 端點統計
    - 認證系統: 2 個端點
    - 承攬商管理: 4 個端點
    - 施工單管理: 6 個端點
    - 簽核管理: 8 個端點
    - 年度資格管理: 6 個端點
    - FaceMatch 整合: 4 個端點
    - 使用者管理: 7 個端點
    - 企業級日誌: 3 個端點
    - **總計: 43 個 API 端點**
    
  version: 2.2.0
  contact:
    name: FaceMatch 技術支援
    email: support@facematch.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5002
    description: TypeScript 開發環境
  - url: http://localhost:5001
    description: Legacy JavaScript 環境
  - url: https://api.facematch.local
    description: 生產環境
paths:
...
  /api/qualifications:
    get:
      tags:
        - 年度資格管理
      summary: 取得年度資格列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: year
          schema:
            type: integer
          description: 篩選年度
          example: 2024
        - in: query
          name: status
          schema:
            type: string
            enum: [ACTIVE, EXPIRED, SUSPENDED]
          description: 篩選狀態
        - in: query
          name: contractorId
          schema:
            type: integer
          description: 承攬商 ID
      responses:
        '200':
          description: 成功取得年度資格列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Qualification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - 年度資格管理
      summary: 建立年度資格
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QualificationCreate'
      responses:
        '201':
          description: 成功建立年度資格
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Qualification'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/qualifications/{id}:
    put:
      tags:
        - 年度資格管理
      summary: 更新年度資格
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 資格 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QualificationUpdate'
      responses:
        '200':
          description: 成功更新年度資格
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Qualification'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - 年度資格管理
      summary: 刪除年度資格
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 資格 ID
      responses:
        '200':
          description: 成功刪除年度資格
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "年度資格已成功刪除"
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/qualifications/{id}/quick-renew:
    post:
      tags:
        - 年度資格管理
      summary: 快速續約
      description: 快速續約到期或即將到期的資格
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 資格 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newExpiryDate
              properties:
                newExpiryDate:
                  type: string
                  format: date
                  description: 新到期日期
                  example: "2025-12-31"
                comment:
                  type: string
                  description: 續約備註
                  example: "年度資格自動續約"
      responses:
        '200':
          description: 快速續約成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Qualification'
                  message:
                    type: string
                    example: "資格已成功續約"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/qualifications/{id}/quick-suspend:
    post:
      tags:
        - 年度資格管理
      summary: 快速暫停
      description: 快速暫停有效的資格
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 資格 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  description: 暫停原因
                  example: "發現安全疑慮，暫停資格"
                suspendUntil:
                  type: string
                  format: date
                  description: 暫停至日期（可選）
                  example: "2024-06-30"
      responses:
        '200':
          description: 快速暫停成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Qualification'
                  message:
                    type: string
                    example: "資格已暫停"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Contractor:
      type: object
      properties:
        id:
          type: integer
          description: 承攬商 ID
          example: 1
        name:
          type: string
          description: 承攬商名稱
          example: "台積電承攬商"
        contact:
          type: string
          description: 聯絡人
          example: "張三"
        phone:
          type: string
          description: 聯絡電話
          example: "02-1234-5678"
        email:
          type: string
          format: email
          description: 電子郵件
          example: "contact@tsmc-contractor.com"
        address:
          type: string
          description: 地址
          example: "新竹科學園區"
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
          description: 狀態
          example: "ACTIVE"
        createdAt:
          type: string
          format: date-time
          description: 建立時間
        updatedAt:
          type: string
          format: date-time
          description: 更新時間
...
    Qualification:
      type: object
      properties:
        id:
          type: integer
          description: 資格 ID
          example: 1
        contractorId:
          type: integer
          description: 承攬商 ID
          example: 1
        contractor:
          $ref: '#/components/schemas/Contractor'
        qualificationType:
          type: string
          description: 資格類型
          example: "安全衛生教育訓練"
        year:
          type: integer
          description: 年度
          example: 2024
        issueDate:
          type: string
          format: date
          description: 發證日期
          example: "2024-01-15"
        expiryDate:
          type: string
          format: date
          description: 到期日期
          example: "2024-12-31"
        status:
          type: string
          enum: [ACTIVE, EXPIRED, SUSPENDED]
          description: 狀態
          example: "ACTIVE"
        certificateNumber:
          type: string
          description: 證書編號
          example: "CERT-2024-001"
        issuingAuthority:
          type: string
          description: 發證機關
          example: "勞動部職業安全衛生署"
        notes:
          type: string
          description: 備註
          example: "年度安全訓練完成"
        createdAt:
          type: string
          format: date-time
          description: 建立時間
        updatedAt:
          type: string
          format: date-time
          description: 更新時間
    QualificationCreate:
      type: object
      required:
        - contractorId
        - qualificationType
        - year
        - issueDate
        - expiryDate
      properties:
        contractorId:
          type: integer
          description: 承攬商 ID
          example: 1
        qualificationType:
          type: string
          description: 資格類型
          example: "安全衛生教育訓練"
        year:
          type: integer
          description: 年度
          example: 2024
        issueDate:
          type: string
          format: date
          description: 發證日期
          example: "2024-01-15"
        expiryDate:
          type: string
          format: date
          description: 到期日期
          example: "2024-12-31"
        certificateNumber:
          type: string
          description: 證書編號
          example: "CERT-2024-001"
        issuingAuthority:
          type: string
          description: 發證機關
          example: "勞動部職業安全衛生署"
        notes:
          type: string
          description: 備註
          example: "年度安全訓練完成"
    QualificationUpdate:
      type: object
      properties:
        qualificationType:
          type: string
          description: 資格類型
        year:
          type: integer
          description: 年度
        issueDate:
          type: string
          format: date
          description: 發證日期
        expiryDate:
          type: string
          format: date
          description: 到期日期
        status:
          type: string
          enum: [ACTIVE, EXPIRED, SUSPENDED]
          description: 狀態
        certificateNumber:
          type: string
          description: 證書編號
        issuingAuthority:
          type: string
          description: 發證機關
        notes:
          type: string
          description: 備註
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: 錯誤代碼
            message:
              type: string
              description: 錯誤訊息
            details:
              type: object
              description: 詳細錯誤資訊
  responses:
    BadRequest:
      description: 請求參數錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "請求參數驗證失敗"
              details:
                field: "name"
                message: "名稱為必填欄位"
    Unauthorized:
      description: 未授權 - 需要登入
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "未授權存取，請先登入"
...
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "請求的資源不存在"
    InternalError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "INTERNAL_ERROR"
              message: "伺服器發生未預期的錯誤"
...